nutritionist-migrated/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
python-dotenv==0.19.0
anthropic==0.18.1
jinja2==3.0.1
aiofiles==0.7.0
pydantic==1.8.2
email-validator==1.1.3 